Belajar-DB/server/requirements.txt
Flask
mysql-connector-python
python-dotenv
flask-cors